Host frontend on selfHostedAws implementation.
