+++ package.json
@@ -59,6 +59,7 @@
     "json5": "~2.2.3",
     "loglevel": "^1.8.1",
     "loglevel-plugin-prefix": "^0.8.4",
+    "mime-types": "^2.1.35",
     "moment": "^2.29.4",
     "mustache": "^4.2.0",
     "node-abort-controller": "^3.1.1",
@@ -96,6 +97,7 @@
     "@types/fs-extra": "^11.0.1",
     "@types/glob": "^8.0.0",
     "@types/jest": "^29.2.2",
+    "@types/mime-types": "^2.1.1",
     "@types/mustache": "^4.2.2",
     "@types/node": "^18.11.4",
     "@types/node-cron": "^3.0.7",
